+++ src/commands/config.ts
@@ -25,8 +25,6 @@ export default class Config extends Command {
         `$ symbol-bootstrap config -p bootstrap`,
         `$ symbol-bootstrap config -p testnet -a dual --password 1234`,
         `$ symbol-bootstrap config -p mainnet -a peer -c custom-preset.yml`,
-        `$ symbol-bootstrap config -p mainnet -a my-custom-assembly.yml -c custom-preset.yml`,
-        `$ symbol-bootstrap config -p my-custom-network.yml -a dual -c custom-preset.yml`,
         `$ echo "$MY_ENV_VAR_PASSWORD" | symbol-bootstrap config -p testnet -a dual`,
     ];
 
@@ -35,17 +33,15 @@ export default class Config extends Command {
         target: CommandUtils.targetFlag,
         password: CommandUtils.passwordFlag,
         noPassword: CommandUtils.noPasswordFlag,
-        preset: flags.string({
+        preset: flags.enum({
             char: 'p',
-            description: `The network preset. It can be provided via custom preset or cli parameter. If not provided, the value is resolved from the target/preset.yml file. Options are: ${Object.keys(
-                Assembly,
-            ).join(', ')}, my-custom-network.yml (advance)`,
+            description: `The network preset. It can be provided via custom preset or cli parameter. If not provided, the value is resolved from the target/preset.yml file.`,
+            options: Object.keys(Preset).map((v) => v as Preset),
         }),
-        assembly: flags.string({
+        assembly: flags.enum({
             char: 'a',
-            description: `The assembly that defines the node(s) layout. It can be provided via custom preset or cli parameter. If not provided, the value is resolved from the target/preset.yml file. Options are: ${Object.keys(
-                Preset,
-            ).join(', ')}, my-custom-assembly.yml (advance)`,
+            description: `The assembly that defines the node(s) layout. It can be provided via custom preset or cli parameter. If not provided, the value is resolved from the target/preset.yml file.`,
+            options: Object.keys(Assembly).map((v) => v as Assembly),
         }),
         customPreset: flags.string({
             char: 'c',