+++ src/commands/start.ts
@@ -26,13 +26,10 @@ export default class Start extends Command {
     static description = 'Single command that aggregates config, compose and run in one line!';
 
     static examples = [
-        `$ symbol-bootstrap start`,
         `$ symbol-bootstrap start -p bootstrap`,
         `$ symbol-bootstrap start -p testnet -a dual`,
         `$ symbol-bootstrap start -p mainnet -a peer -c custom-preset.yml`,
         `$ symbol-bootstrap start -p testnet -a dual --password 1234`,
-        `$ symbol-bootstrap start -p mainnet -a my-custom-assembly.yml -c custom-preset.yml`,
-        `$ symbol-bootstrap start -p my-custom-network.yml -a dual -c custom-preset.yml`,
         `$ echo "$MY_ENV_VAR_PASSWORD" | symbol-bootstrap start -p testnet -a dual`,
     ];